Move BaseVisualizer to its own file.
